+++ acceptance/tests/partitions/main_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	testsuite "github.com/hashicorp/consul-k8s/charts/consul/test/acceptance/framework/suite"
+	testsuite "github.com/hashicorp/consul-k8s/acceptance/framework/suite"
 )
 
 var suite testsuite.Suite