+++ control-plane/subcommand/server-acl-init/rules.go
@@ -7,6 +7,8 @@ import (
 )
 
 type rulesData struct {
+	EnablePartitions        bool
+	PartitionName           string
 	EnableNamespaces        bool
 	SyncConsulDestNS        string
 	SyncEnableNSMirroring   bool
@@ -34,29 +36,62 @@ service "consul-snapshot" {
    policy = "write"
 }`
 
+// The enterprise license rules are acl="write" inside partitions as operator="write"
+// is unsupported in partitions.
 const entLicenseRules = `operator = "write"`
+const entPartitionLicenseRules = `acl = "write"`
 
-const crossNamespaceRules = `namespace_prefix "" {
-  service_prefix "" {
-    policy = "read"
+// The partition token is utilized by the partition-init job and server-acl-init in
+// non-default partitions. This token requires permissions to create partitions, read the
+// agent endpoint during startup and have the ability to create an auth-method within a namespace
+// for any partition.
+const partitionRules = `operator = "write"
+agent_prefix "" {
+  policy = "read"
+}
+partition_prefix "" {
+  namespace_prefix "" {
+    acl = "write"
   }
-  node_prefix "" {
-    policy = "read"
+}`
+
+func (c *Command) crossNamespaceRules() (string, error) {
+	crossNamespaceRulesTpl := `{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
+{{- end }}
+  namespace_prefix "" {
+    service_prefix "" {
+      policy = "read"
+    }
+    node_prefix "" {
+      policy = "read"
+    }
   }
-} `
+{{- if .EnablePartitions }}
+}
+{{- end }}`
+
+	return c.renderRules(crossNamespaceRulesTpl)
+}
 
 func (c *Command) agentRules() (string, error) {
 	agentRulesTpl := `
+{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
+{{- end }}
   node_prefix "" {
     policy = "write"
   }
 {{- if .EnableNamespaces }}
-namespace_prefix "" {
+  namespace_prefix "" {
 {{- end }}
-  service_prefix "" {
-    policy = "read"
-  }
+    service_prefix "" {
+      policy = "read"
+    }
 {{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
@@ -80,16 +115,22 @@ func (c *Command) anonymousTokenRules() (string, error) {
 	// ACL token. Thus the anonymous policy must
 	// allow reading all services.
 	anonTokenRulesTpl := `
-{{- if .EnableNamespaces }}
-namespace_prefix "" {
+{{- if .EnablePartitions }}
+partition_prefix "" {
 {{- end }}
-  node_prefix "" {
-     policy = "read"
-  }
-  service_prefix "" {
-     policy = "read"
-  }
 {{- if .EnableNamespaces }}
+  namespace_prefix "" {
+{{- end }}
+    node_prefix "" {
+       policy = "read"
+    }
+    service_prefix "" {
+       policy = "read"
+    }
+{{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
@@ -133,19 +174,25 @@ namespace_prefix "" {
 
 func (c *Command) ingressGatewayRules(name, namespace string) (string, error) {
 	ingressGatewayRulesTpl := `
-{{- if .EnableNamespaces }}
-namespace "{{ .GatewayNamespace }}" {
+{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
 {{- end }}
-  service "{{ .GatewayName }}" {
-     policy = "write"
-  }
-  node_prefix "" {
-    policy = "read"
-  }
-  service_prefix "" {
-    policy = "read"
-  }
 {{- if .EnableNamespaces }}
+  namespace "{{ .GatewayNamespace }}" {
+{{- end }}
+    service "{{ .GatewayName }}" {
+       policy = "write"
+    }
+    node_prefix "" {
+      policy = "read"
+    }
+    service_prefix "" {
+      policy = "read"
+    }
+{{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
@@ -159,16 +206,22 @@ namespace "{{ .GatewayNamespace }}" {
 // of the initial implementation
 func (c *Command) terminatingGatewayRules(name, namespace string) (string, error) {
 	terminatingGatewayRulesTpl := `
-{{- if .EnableNamespaces }}
-namespace "{{ .GatewayNamespace }}" {
+{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
 {{- end }}
-  service "{{ .GatewayName }}" {
-     policy = "write"
-  }
-  node_prefix "" {
-    policy = "read"
-  }
 {{- if .EnableNamespaces }}
+  namespace "{{ .GatewayNamespace }}" {
+{{- end }}
+    service "{{ .GatewayName }}" {
+       policy = "write"
+    }
+    node_prefix "" {
+      policy = "read"
+    }
+{{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
@@ -207,22 +260,33 @@ func (c *Command) injectRules() (string, error) {
 	// The Connect injector needs permissions to create namespaces when namespaces are enabled.
 	// It must also create/update service health checks via the endpoints controller.
 	// When ACLs are enabled, the endpoints controller needs "acl:write" permissions
-	// to delete ACL tokens created via "consul login".
+	// to delete ACL tokens created via "consul login". policy = "write" is required when
+	// creating namespaces within a partition.
 	injectRulesTpl := `
+{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
+{{- else }}
 {{- if .EnableNamespaces }}
-operator = "write"
+  operator = "write"
 {{- end }}
-node_prefix "" {
-  policy = "write"
-}
-{{- if .EnableNamespaces }}
-namespace_prefix "" {
 {{- end }}
-  acl = "write"
-  service_prefix "" {
+  node_prefix "" {
     policy = "write"
   }
 {{- if .EnableNamespaces }}
+  namespace_prefix "" {
+{{- end }}
+{{- if .EnablePartitions }}
+    policy = "write"
+{{- end }}
+    acl = "write"
+    service_prefix "" {
+      policy = "write"
+    }
+{{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}`
 	return c.renderRules(injectRulesTpl)
@@ -237,43 +301,62 @@ func (c *Command) aclReplicationRules() (string, error) {
 	// datacenters during federation since in order to start ACL replication,
 	// we need a token with both replication and agent permissions.
 	aclReplicationRulesTpl := `
-operator = "write"
-agent_prefix "" {
-  policy = "read"
-}
-node_prefix "" {
-  policy = "write"
-}
-{{- if .EnableNamespaces }}
-namespace_prefix "" {
+{{- if .EnablePartitions }}
+partition "default" {
 {{- end }}
-  acl = "write"
-  service_prefix "" {
+  operator = "write"
+  agent_prefix "" {
     policy = "read"
-    intentions = "read"
+  }
+  node_prefix "" {
+    policy = "write"
   }
 {{- if .EnableNamespaces }}
+  namespace_prefix "" {
+{{- end }}
+    acl = "write"
+    service_prefix "" {
+      policy = "read"
+      intentions = "read"
+    }
+{{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
 	return c.renderRules(aclReplicationRulesTpl)
 }
 
+// policy = "write" is required when creating namespaces within a partition.
 func (c *Command) controllerRules() (string, error) {
 	controllerRules := `
-operator = "write"
+{{- if .EnablePartitions }}
+partition "{{ .PartitionName }}" {
+  mesh = "write"
+  acl = "write"
+{{- else }}
+  operator = "write"
+{{- end }}
 {{- if .EnableNamespaces }}
 {{- if .InjectEnableNSMirroring }}
-namespace_prefix "{{ .InjectNSMirroringPrefix }}" {
+  namespace_prefix "{{ .InjectNSMirroringPrefix }}" {
 {{- else }}
-namespace "{{ .InjectConsulDestNS }}" {
+  namespace "{{ .InjectConsulDestNS }}" {
 {{- end }}
 {{- end }}
-  service_prefix "" {
+{{- if .EnablePartitions }}
     policy = "write"
-    intentions = "write"
-  }
+{{- end }}
+    service_prefix "" {
+      policy = "write"
+      intentions = "write"
+    }
 {{- if .EnableNamespaces }}
+  }
+{{- end }}
+{{- if .EnablePartitions }}
 }
 {{- end }}
 `
@@ -282,6 +365,8 @@ namespace "{{ .InjectConsulDestNS }}" {
 
 func (c *Command) rulesData() rulesData {
 	return rulesData{
+		EnablePartitions:        c.flagEnablePartitions,
+		PartitionName:           c.flagPartitionName,
 		EnableNamespaces:        c.flagEnableNamespaces,
 		SyncConsulDestNS:        c.flagConsulSyncDestinationNamespace,
 		SyncEnableNSMirroring:   c.flagEnableSyncK8SNSMirroring,